Small refactor, moving chunk related files into a module.
