+++ feather/common/src/lib.rs
@@ -19,17 +19,11 @@ pub use window::Window;
 
 pub mod events;
 
-pub mod chunk_worker;
+pub mod chunk;
 mod region_worker;
-
-pub mod chunk_cache;
 
 pub mod world;
 pub use world::World;
-
-mod chunk_loading;
-
-mod chunk_entities;
 
 pub mod chat;
 pub use chat::ChatBox;
@@ -41,8 +35,8 @@ pub mod interactable;
 /// Registers gameplay systems with the given `Game` and `SystemExecutor`.
 pub fn register(game: &mut Game, systems: &mut SystemExecutor<Game>) {
     view::register(game, systems);
-    chunk_loading::register(game, systems);
-    chunk_entities::register(systems);
+    chunk::loading::register(game, systems);
+    chunk::entities::register(systems);
     interactable::register(game);
 
     game.add_entity_spawn_callback(entities::add_entity_components);
